Fix DXVR runtime compatibility
The DXVR runtime sends the XR_TYPE_EVENT_DATA_INTERACTION_PROFILE_CHANGED event much earlier in the startup process, before the interaction profile has had a chance to be queried.  I moved the querying of the interaction profile to the start of PumpEvents to ensure that it is always ready before the event comes.